+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities
 
+## [1.0.1] - 2021.01.19
+
+### Changed
+
+- Updated vendor dependencies.
+- Updated HTTP status code to return `200` instead of `201` on successful authentication.
+- Renamed `Webhooks` controller to `PublicController` and added a public "API status" route.
+
 ## [1.0.0] - 2020.12.05
 
 ### Added